+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add support for TPROXY rules
+- `Set::flush`
+- Implement `SetKey` for `u16`
+
 
 ## [0.9.4] - 2026-08-11
 ### Added